refactor: move Provider tests to forge_infra and ensure sequential execution
This commit moves all provider-related tests from forge_domain/src/provider.rs
to forge_infra/src/env.rs which is a more appropriate location for environment
variable testing. Also adds the serial_test crate to ensure tests run sequentially
since they modify shared environment variables.

Changes:
- Add serial_test as a dev dependency to forge_infra
- Move all provider tests to env.rs
- Add #[serial] attribute to tests to prevent race conditions
- Remove tests from provider.rs
