+++ crates/forge_infra/Cargo.toml
@@ -15,3 +15,6 @@ serde_json = "1.0.138"
 qdrant-client = "1.13.0"
 reqwest = {version = "0.12.12", features = ["json", "rustls-tls"], default-features = false}
 serde = { version = "1.0", features = ["derive"] }
+
+[dev-dependencies]
+serial_test = "2.0.0"
