+++ crates/forge_infra/src/env.rs
@@ -65,3 +65,110 @@ impl EnvironmentService for ForgeEnvironmentService {
         self.get()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use std::env;
+    use serial_test::serial;
+    use forge_domain::Provider;
+
+    // reset the env variables for reliable tests
+    fn reset_env() {
+        env::remove_var("FORGE_KEY");
+        env::remove_var("FORGE_PROVIDER_URL");
+        env::remove_var("OPEN_ROUTER_KEY");
+        env::remove_var("OPENAI_API_KEY");
+        env::remove_var("ANTHROPIC_API_KEY");
+    }
+
+    #[test]
+    #[serial]
+    fn test_provider_from_env_with_forge_key_and_without_provider_url() {
+        reset_env();
+        env::set_var("FORGE_KEY", "some_forge_key");
+
+        let provider = Provider::from_env();
+        assert_eq!(provider, None);
+    }
+
+    #[test]
+    #[serial]
+    fn test_provider_from_env_with_forge_key() {
+        reset_env();
+        env::set_var("FORGE_KEY", "some_forge_key");
+        env::set_var("FORGE_PROVIDER_URL", "https://api.openai.com/v1/");
+
+        let provider = Provider::from_env();
+        assert_eq!(provider, Some(Provider::OpenAI));
+    }
+
+    #[test]
+    #[serial]
+    fn test_provider_from_env_with_open_router_key() {
+        reset_env();
+        env::set_var("OPEN_ROUTER_KEY", "some_open_router_key");
+
+        let provider = Provider::from_env();
+        assert_eq!(provider, Some(Provider::OpenRouter));
+    }
+
+    #[test]
+    #[serial]
+    fn test_provider_from_env_with_openai_key() {
+        reset_env();
+        env::set_var("OPENAI_API_KEY", "some_openai_key");
+
+        let provider = Provider::from_env();
+        assert_eq!(provider, Some(Provider::OpenAI));
+    }
+
+    #[test]
+    #[serial]
+    fn test_provider_from_env_with_anthropic_key() {
+        reset_env();
+        env::set_var("ANTHROPIC_API_KEY", "some_anthropic_key");
+
+        let provider = Provider::from_env();
+        assert_eq!(provider, Some(Provider::Anthropic));
+    }
+
+    #[test]
+    #[serial]
+    fn test_provider_from_env_with_no_keys() {
+        reset_env();
+        let provider = Provider::from_env();
+        assert_eq!(provider, None);
+    }
+
+    #[test]
+    #[serial]
+    fn test_from_url() {
+        assert_eq!(
+            Provider::from_url("https://api.openai.com/v1/"),
+            Some(Provider::OpenAI)
+        );
+        assert_eq!(
+            Provider::from_url("https://api.openrouter.io/v1/"),
+            Some(Provider::OpenRouter)
+        );
+        assert_eq!(
+            Provider::from_url("https://api.anthropic.com/v1/"),
+            Some(Provider::Anthropic)
+        );
+        assert_eq!(Provider::from_url("https://unknown.url/"), None);
+    }
+
+    #[test]
+    #[serial]
+    fn test_to_url() {
+        assert_eq!(Provider::OpenAI.to_base_url(), "https://api.openai.com/v1/");
+        assert_eq!(
+            Provider::OpenRouter.to_base_url(),
+            "https://api.openrouter.io/v1/"
+        );
+        assert_eq!(
+            Provider::Anthropic.to_base_url(),
+            "https://api.anthropic.com/v1/"
+        );
+    }
+}
